move closer into own test, add gaggle version
