+++ tests/one_taskset.rs
@@ -5,7 +5,6 @@ mod common;
 
 use goose::goose::GooseMethod;
 use goose::prelude::*;
-use std::sync::Arc;
 use std::thread;
 
 const INDEX_PATH: &str = "/";
@@ -213,155 +212,6 @@ fn test_single_taskset_gaggle() {
 
     // Verify that Goose started the correct number of users.
     assert!(goose_metrics.users == manager_configuration.users.unwrap());
-}
-
-#[test]
-// Load test with a single task set containing two weighted tasks setup via closure.
-// Validate weighting and statistics.
-fn test_single_taskset_closure() {
-    // @todo Move out to common.rs.
-    #[derive(Debug)]
-    struct LoadtestEndpoint<'a> {
-        pub path: &'a str,
-        pub status_code: u16,
-        pub weight: usize,
-    };
-
-    // Configure endpoints to test.
-    let test_endpoints = vec![
-        LoadtestEndpoint {
-            path: INDEX_PATH,
-            status_code: 200,
-            weight: 9,
-        },
-        LoadtestEndpoint {
-            path: ABOUT_PATH,
-            status_code: 200,
-            weight: 3,
-        },
-    ];
-
-    // Start mock server.
-    let server = MockServer::start();
-
-    // Build configuration.
-    let config = common::build_configuration(
-        &server,
-        vec![
-            "--no-reset-metrics",
-            "--no-task-metrics",
-            "--status-codes",
-            "--users",
-            &test_endpoints.len().to_string(),
-            "--hatch-rate",
-            &(2 * test_endpoints.len()).to_string(),
-        ],
-    );
-
-    // Setup mock endpoints.
-    let mut mock_endpoints = Vec::with_capacity(test_endpoints.len());
-    for (idx, item) in test_endpoints.iter().enumerate() {
-        let path = item.path;
-        let mock_endpoint = Mock::new()
-            .expect_method(GET)
-            .expect_path(path)
-            .return_status(item.status_code.into())
-            .create_on(&server);
-
-        // Ensure the index matches.
-        assert!(idx == mock_endpoints.len());
-        mock_endpoints.push(mock_endpoint);
-    }
-
-    // Build dynamic taskset.
-    let mut taskset = GooseTaskSet::new("LoadTest");
-    for item in &test_endpoints {
-        let path = item.path;
-        let weight = item.weight;
-
-        let closure: GooseTaskFunction = Arc::new(move |user| {
-            Box::pin(async move {
-                let _goose = user.get(path).await?;
-
-                Ok(())
-            })
-        });
-
-        let task = GooseTask::new(closure).set_weight(weight).unwrap();
-        // We need to do the variable dance as taskset.register_task returns self and hence moves
-        // self out of `taskset`. By storing it in a new local variable and then moving it over
-        // we can avoid that error.
-        let new_taskset = taskset.register_task(task);
-        taskset = new_taskset;
-    }
-
-    // Run the loadtest.
-    let goose_metrics = crate::GooseAttack::initialize_with_config(config.clone())
-        .unwrap()
-        .register_taskset(taskset)
-        .execute()
-        .unwrap();
-
-    // Ensure that the right paths have been called.
-    for (idx, item) in test_endpoints.iter().enumerate() {
-        let mock_endpoint = &mock_endpoints[idx];
-
-        let format_item = |message, assert_item| {
-            return format!("{} for item = {:#?}", message, assert_item);
-        };
-
-        // Confirm that we loaded the mock endpoint.
-        assert!(
-            mock_endpoint.times_called() > 0,
-            format_item("Endpoint was not called > 0", &item)
-        );
-        let expect_error = format_item("Item does not exist in goose_metrics", &item);
-        let endpoint_metrics = goose_metrics
-            .requests
-            .get(&format!("GET {}", item.path))
-            .expect(&expect_error);
-
-        assert!(
-            endpoint_metrics.path == item.path,
-            format_item(
-                &format!("{} != {}", endpoint_metrics.path, item.path),
-                &item
-            )
-        );
-        assert!(endpoint_metrics.method == GooseMethod::GET);
-
-        // Confirm that Goose and the server saw the same number of page loads.
-        let status_code: u16 = item.status_code;
-
-        assert!(
-            endpoint_metrics.response_time_counter == mock_endpoint.times_called(),
-            format_item("response_time_counter != times_called()", &item)
-        );
-        assert!(
-            endpoint_metrics.status_code_counts[&status_code] == mock_endpoint.times_called(),
-            format_item("status_code_counts != times_called()", &item)
-        );
-        assert!(
-            endpoint_metrics.success_count == mock_endpoint.times_called(),
-            format_item("success_count != times_called()", &item)
-        );
-        assert!(
-            endpoint_metrics.fail_count == 0,
-            format_item("fail_count != 0", &item)
-        );
-    }
-
-    // Test specific things directly access the mock endpoints here.
-    let index = &mock_endpoints[0];
-    let about = &mock_endpoints[1];
-
-    // Confirm that we loaded the index roughly three times as much as the about page.
-    let one_third_index = index.times_called() / 3;
-    let difference = about.times_called() as i32 - one_third_index as i32;
-    assert!(difference >= -2 && difference <= 2);
-
-    // Verify that Goose started the correct number of users.
-    assert!(goose_metrics.users == config.users.unwrap());
 }
 
 #[test]
